+++ package/config/config.js
@@ -1,5 +1,4 @@
 
-const { Sequelize } = require("sequelize");
 const { argv } = require("../utils/argv");
 
 const config = {
@@ -14,33 +13,4 @@ const config = {
   port: argv.p || argv.port || 5432
 }
 
-function fillUpConfigObj(option) {
-  console.log('HEREEEEEEEEEEEEEEEEEEEEEEE');
-  for (const key in config) {
-    if (option[key]) {
-      config[key] = option[key];
-    }
-  }
-  // return config;
-  console.log(config);
-  const sequelize = new Sequelize(
-    config.database,
-    config.username,
-    config.password,
-    {
-      host: config.host,
-      dialect: config.dialect,
-      port: config.port,
-      logging: config.logging,
-    }
-  );
-
-  return sequelize
-}
-
-module.exports = {fillUpConfigObj} 
-
-
-
-
-// module.exports = config
+module.exports = config